Additional modifications for Next13 project setup
