+++ app/_assets/styles/tailwind.css
