+++ app/_components/forms/PatientTable.jsx
@@ -1,4 +1,3 @@
-import { Container } from '@/components/layout/Container'
 
 const patient = [
   { number: '1854', familyName: 'Lavelle', givenName: 'Tim', age: '41' },
