+++ app/_utils/displayProvinces.js
@@ -1,28 +1,28 @@
-import {React, useMemo} from 'react';
-import GetProvincesAPI from '@/utils/services/GetProvincesAPI';
-import {Field, useFormikContext } from 'formik';
-import {useTranslation} from 'next-i18next';
+import { React, useMemo } from 'react';
+import GetProvincesAPI from '@/_utils/services/GetProvincesAPI';
+import { Field, useFormikContext } from 'formik';
+import { useTranslation } from 'next-i18next';
 
-export default function DisplayProvinces(props){
-  
-  const [ provinces ] = GetProvincesAPI();
-  const {setFieldValue, errors, touched} = useFormikContext();
+export default function DisplayProvinces(props) {
+
+  const [provinces] = GetProvincesAPI();
+  const { setFieldValue, errors, touched } = useFormikContext();
   const { t } = useTranslation();
-  
-  function handleSelectedProvince(pr){
+
+  function handleSelectedProvince(pr) {
     console.log('Fired')
-    console.log('Selected Province = ' +pr)
+    console.log('Selected Province = ' + pr)
     // Want to send 'pr' to displayDistricts.js in order to pull the appropriate Districts
-    props.callBack({id: pr})
+    props.callBack({ id: pr })
   }
-  
+
   return (
     <>
       <Field
         as="select"
         id="pxProvince"
         name="pxProvince"
-        onChange={(e)=> {
+        onChange={(e) => {
           handleSelectedProvince(e.target.value)
           setFieldValue('pxProvince', e.target.value)
         }}
