+++ next.config.js
@@ -10,7 +10,15 @@ const nextConfig = {
   experimental: {
     newNextLinkBehavior: true
   },
-  trailingSlash: false
+  trailingSlash: false,
+  images: {
+    remotePatterns: [
+      {
+        protocol: 'https',
+        hostname: 'images.unsplash.com'
+      }
+    ]
+  }
 }
 
 module.exports = nextConfig;
